pyproject.toml
[tool.poetry]
name = "imaginex-lambda"
version = "0.7.0"
description = "Image optimizer for processing NextJS image requests"
authors = ["Jan Soukup <jan@ssoukup.com>"]

[tool.poetry.dependencies]
python = "~3.8"
pillow = "^9.3.0"
requests = "^2.28.1"
botocore = "^1.29.41"
filetype = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
python-dotenv = "^1.0.0"
pandas = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
